+++ src/js/mixin/internal/animate-fade.js
@@ -1,10 +1,10 @@
 import {getRows} from '../../core/margin';
-import {addClass, children, css, hasClass, height, isInView, once, Promise, removeClass, sortBy, toNumber, Transition} from 'uikit-util';
+import {addClass, children, css, fastdom, hasClass, height, isInView, once, Promise, removeClass, sortBy, toNumber, toWindow, Transition, trigger} from 'uikit-util';
 
 const clsLeave = 'uk-transition-leave';
 const clsEnter = 'uk-transition-enter';
 
-export default function fade(action, target, duration, stagger = 40) {
+export default function fade(action, target, duration, stagger = 0) {
 
     const index = transitionIndex(target, true);
     const propsIn = {opacity: 1};
@@ -32,9 +32,14 @@ export default function fade(action, target, duration, stagger = 40) {
         addClass(target, clsEnter);
         action();
 
+        // Force update
+        trigger(toWindow(target), 'resize');
+        fastdom.flush();
+
         css(children(target), {opacity: 0});
 
-        return new Promise(resolve => {
+        // Two frames to ensure UIkit updates propagated
+        return new Promise(resolve =>
             requestAnimationFrame(() => {
 
                 const nodes = children(target);
@@ -64,8 +69,8 @@ export default function fade(action, target, duration, stagger = 40) {
                     }
                     resolve();
                 });
-            });
-        });
+            })
+        );
     });
 
     return hasClass(target, clsLeave)