+++ src/js/mixin/slider-drag.js
@@ -1,4 +1,4 @@
-import {closest, css, getEventPos, hasTouch, includes, isRtl, isTouch, off, on, pointerCancel, pointerDown, pointerMove, pointerUp, selInput, trigger} from 'uikit-util';
+import {closest, css, getEventPos, includes, isRtl, isTouch, off, on, pointerCancel, pointerDown, pointerMove, pointerUp, selInput, trigger} from 'uikit-util';
 
 export default {
 
@@ -88,7 +88,10 @@ export default {
                 this.prevIndex = this.index;
             }
 
-            // See above workaround notice
+            // Workaround for iOS's inert scrolling preventing pointerdown event
+            // https://developer.mozilla.org/en-US/docs/Web/CSS/touch-action
+            on(this.list, 'touchmove', this.move, {passive: false});
+
             on(document, pointerMove, this.move, {passive: false});
             on(document, `${pointerUp} ${pointerCancel}`, this.end, true);
 
@@ -104,11 +107,7 @@ export default {
                 return;
             }
 
-            // Workaround for iOS's inert scrolling preventing pointerdown event
-            // https://developer.mozilla.org/en-US/docs/Web/CSS/touch-action
-            if (!this.dragging && hasTouch) {
-                on(this.list, 'touchmove', preventDefault, {passive: false});
-            }
+            e.cancelable && e.preventDefault();
 
             this.dragging = true;
             this.dir = (distance < 0 ? 1 : -1);
@@ -172,9 +171,9 @@ export default {
 
         end() {
 
+            off(this.list, 'touchmove', this.move, {passive: false});
             off(document, pointerMove, this.move, {passive: false});
             off(document, `${pointerUp} ${pointerCancel}`, this.end, true);
-            off(this.list, 'touchmove', preventDefault, {passive: false});
 
             if (this.dragging) {
 
@@ -213,8 +212,4 @@ export default {
 
 function hasTextNodesOnly(el) {
     return !el.children.length && el.childNodes.length;
-}
-
-function preventDefault(e) {
-    e.cancelable && e.preventDefault();
 }